🔐 : – Prefer secure server tokens for server joins
What: add secure token resolver, update discovery logic, refresh tests.

Why: prevent server joins from reusing agent/node tokens.

How to test: pytest tests/scripts/test_k3s_discover_token_resolution.py
